Repro gen conflict in COMPUTE_STATUS_DISABLED handling
The COMPUTE_STATUS_DISABLED trait is supposed to be added to the compute
RP when the compute service is disabled, and the trait supposed to be
removed when the service is enabled again. However adding and removing
traits is subject to generation conflict in placement. The original
implementation of blueprint pre-filter-disabled-computes noticed this
and prints a detailed warning message while the API operation succeeds.
We can ignore the conflict this way because the periodic
update_available_resource() call will re-sync the traits later.

Still this gives human noticeable time window where the trait and the
service state are not in sync.

Setting the compute service disable is the smaller problem as the
scheduler still uses the ComputeFilter that filters the computes based
on the service api. So during the enable ->disable race window we only
lose scheduling performance as the placement filter is inefficient.

In case of setting the compute service to enabled the race is more
visible as the placement pre_filter will filter out the compute that
is enable by the admin until the re-sync happens. If the de-sync would
only happen due to high load on the given compute the such delay could
be explained by the load itself. However de-sync can happen simply due
to a new instance boot on the compute.

This patch adds a functional test that reproduce the original problem.

Related-Bug: #1886418

Change-Id: Ib980b1ba68ffcfe51a15dce10eb9f42ef12d7260