func tests: move _run_periodics() into base class
There are two almost identical implementations of the _run_periodics()
helper - and a third one would have joined them in a subsequent patch,
if not for this patch. This patch moves the _run_periodics() to the
base test class. In addition, _run_periodics() depends on the
self.computes dict used for compute service tracking. The method that
populates that dict, _start_compute(), is therefore also moved to the
base class.

This enables some light refactoring of existing tests that need
either the _run_periodics() helper, or the compute service tracking.

In addition, a needless override of _start_compute() in
test_aggregates that provided no added value is removed. This is done
to avoid any potential confusion around _start_compute()'s role.

Conflicts:
  nova/tests/functional/compute/test_cache_image.py
  nova/tests/functional/integrated_helpers.py
  nova/tests/functional/test_scheduler.py

NOTE(stephenfin): Conflicts are due to the two test files not existing
yet and the large number of helpers missing from the
'integrated_helpers' class.

Changes:
  nova/tests/functional/regressions/test_bug_1889108.py

NOTE(stephenfin): Changes are due to a backport that was modified to
remove use of the functions we're adding here.

Change-Id: I33d8ac0a1cae0b2d275a21287d5e44c008a68122
(cherry picked from commit ee05cd8b9e73f8a9a43dbe9ce028e6a6beaec2bd)