func tests: move _run_periodics() into base class
There are two almost identical implementations of the _run_periodics()
helper - and a third one would have joined them in a subsequent patch,
if not for this patch. This patch moves the _run_periodics() to the
base test class. In addition, _run_periodics() depends on the
self.computes dict used for compute service tracking. The method that
populates that dict, _start_compute(), is therefore also moved to the
base class.

This enables some light refactoring of existing tests that need
either the _run_periodics() helper, or the compute service tracking.

In addition, a needless override of _start_compute() in
test_aggregates that provided no added value is removed. This is done
to avoid any potential confusion around _start_compute()'s role.

Change-Id: I36dd64dc272ea1743995b3b696323a9431666489

safdasdf

Change-Id: I33d8ac0a1cae0b2d275a21287d5e44c008a68122