tests: work around malformed serial XML
The serial elements in fakelibvirt's XMLDesc() are malformed. They are
only enough to pass existing unit tests, but are not future-proof, and
would not work with something like migration._update_serial_xml(). In
addition, two of them (type=file and type=pty) are not actually used by
anything.

This patch removes the unused types and starts checking the
CONF.serial_console.enabled flag before including any serial XML in
the fake domain. This allows tests that don't set
serial_console_enabled (like the subsequent NUMA LM func test) to not
have to deal with the broken serial XML, while tests that relied on it
in the past now just set serial_console_enabled = True.

Change-Id: Ie74fcd689a627c178f80ec60df2dfef40576f182