conf: Deprecate '[scheduler] driver'
With the removal of 'CachingScheduler' in Stein (19.0.0), the
'FilterScheduler' is the only in-tree scheduler provided by nova. The
removal of the caching scheduler was motivated in good part because
"more and more functionality within nova [now] relies on managing
(sometimes complex) resource allocations in the placement service". This
same argument applies for out-of-tree drivers. It's unlikely that there
are users with out-of-tree drivers that have good placement integration
and significantly outperform the default filter scheduler. As a result,
it's probably time to close this extension point in favour of continued
improvements on the filter scheduler and placement along with custom
filters and weighters for the filter scheduler.

Change-Id: Id65e0b16fe062509097b8e3f58eb81bfe1cc2bfa
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>