images: Remove Libvirt specific configurable use from qemu_img_info
This change addresses an old TODO in the images module by dropping the
use of a Libvirt specific configurable from the qemu_img_info function.
We can identify RBD based volumes by checking for 'rbd:' at the start of
the path provided to the function instead of using the configurable.

Change-Id: Ife9e67d5c71f4cca825dff713f54ec955508f6e6