Avoid circular reference during serialization
When an instance with numa topology is re-scheduled the conductor
migrate task blows with circular reference during request spec
serialization. It happens because there are ovos in the request spec
that jsonutils.dumps only serialize if requested explicitly.

This patch makes the explicit request.

This is a stable only bug fix as the borken code was removed in Stein by
the feature patch I4244f7dd8fe74565180f73684678027067b4506e

Closes-Bug: #1864665

Change-Id: I1942bfa9bd1baf8738d34c287216db7b59000a36