Switching new default roles in os-volumes-attachments policies
We are introducing new default roles with scope capability
in policies, details as bellow:

  - For list and show volume attachment APIs change RULE_ADMIN_OR_OWNER
    to PROJECT_READER_OR_SYSTEM_READER role.
  - For create and delete attachment APIs change RULE_ADMIN_OR_OWNER to
    PROJECT_MEMBER_OR_SYSTEM_ADMIN role.
  - For update volume attachment API changes RULE_ADMIN_API to
    SYSTEM_ADMIN role.

And this simulates the future where we drop the deprecation
fall back in the policy by overriding the rules with a version where
there are no deprecated rule options. Operators can do the same by
adding overrides in their policy files that match the default but
stop the rule deprecation fallback from happening.

Partial implement blueprint policy-defaults-refresh

Change-Id: I21aed9418fbf5d27717a3eed41ec4a0c0a523d6b