libvirt: Remove MIN_LIBVIRT_MULTIATTACH
The updated minimum required libvirt (4.0.0) and QEMU (2.11)
for "Ussuri" satisfy the version requirements; this was done
in Change-Id: Ia18e9be4d (22c1916b49 — libvirt: Bump
MIN_{LIBVIRT,QEMU}_VERSION for "Ussuri", 2019-11-19).

Drop the version constant MIN_LIBVIRT_MULTIATTACH and now-needless
compatibility code; adjust/remove tests.

Note that the modified code is a little unusual, in that it was
checking that QEMU's version was *less* than X or libvirt's version was
*greater* than Y, as opposed to both being lesser or greater than X.
This was because we were working around [1], whereby QEMU 2.10.0 added
an image locking feature that broke multiattach, while libvirt 3.10.0
added support for a shareable flag to pass the correct flags to QEMU
and fix the feature.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1378242

Change-Id: Idab5df9d1b733a7df6bbcaf44aa01e444ad4099d
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>