Add a tests to check when legacy access is removed
This test is trying to check how an operator is able to adopt things
like project_reader. We need project_member APIs to not be available
to the project_reader.

This basically simulates the future where we drop the deprecation
fall back in the policy by overriding the rules with a version where
there are no deprecated rule options. Operators can do the same by
adding overrides in their policy files that match the default but
stop the rule deprecation fallback from happening.

Partial implement blueprint policy-defaults-refresh

Change-Id: Ic6e1fb081fe61262f718efc045b3b3f28694e8a7