Deprecate the vmwareapi driver
As of now, questions on the mailing list are going unanswered, and the
Nova team does not have a clear representative owner for the driver
to which bugs and other reports can be directed. There does not appear
to be a CI system running tests for the driver anymore, and the latest
indication from the community[1] points to it being potentially broken with
devstack.

This patch starts the deprecation timer for the driver and/or serves
as a flare to gauge interest (or lack thereof) in continuing to maintain
the driver.

1: http://lists.openstack.org/pipermail/openstack-discuss/2020-March/013066.html

Change-Id: Ie39e9605dc8cebff3795a29ea91dc08ee64a21eb