Make serialize_args handle exception messages safely
We have recently move towards handling exceptions that may be viewed
by API users in a way that results in either the exc.format_message()
text, or just the exception class name (without details) in order to
avoid leaking sensitive information out of the API. This makes the
serialize_args decorator behave that way. Note that because for a
remotable method, serialize_args gets called twice (once on the remote
side before the call and again on the local side at replay), which
means we need to gate that message-or-class-name behavior on whether
or not we are seeing a true Exception object.

Change-Id: I22fb253616443766ef21692735b80faa284fa26d