Fix intermittently failing regression case
The test_unshelve_offloaded_fails_due_to_neutron could fail due to race
condition. The test case only waits for the first instance.save() call
at [1] but the allocation delete happens after it. This causes that the
test case can still see the allocation of the offloaded server in
placement.

The fix makes sure that the test waits for the second instance.save() by
checking for the host of the instance.

[1] https://github.com/openstack/nova/blob/stable/rocky/nova/compute/manager.py#L5274-L5288

Related-Bug #1862633

Change-Id: Ic1c3d35749fbdc7f5b6f6ec1e16b8fcf37c10de8