Refine and introduce correct parameters for test_get_guest_config_numa_host_instance_topo_cpu_pinning
'sockets_per_cell', 'cores_per_socket' and 'threads_per_core'
are not the initializeing arguments of 'NUMATopology' object.

In test test_get_guest_config_numa_host_instance_topo_cpu_pinning
the code path does not care about the correctness of a host
numa toplogy due to being an instance with 'DEDICATED' cpu
policy, but We'd better refine it and create a correct numa
topology to avoid misunderstanding of code.

Change-Id: Ic1c9cf16c482939d0761d0cdab66c8eac07cad7b