Update quota apis with keystone limits and usage
This makes use of the keystone APIs to get limits from Keystone when
showing the user the limits on their project.

Note we also change the default in_use amount from -1, which is what the
no op driver originally used, to 0, which matches what the db driver
typically returns for deprecated quota values, like floating ip limits.
This seems a more sane value to respond with, given we don't count the
usage for those values.

blueprint unified-limits-nova

Change-Id: I933dc135a364b14ddadc8eee67b42d8e1278a9ae