hardware: Don't consider overhead CPUs for unpinned instances
The 'cpuset_reserved' argument for the '_numa_fit_instance_cell'
function is only ever non-zero if the 'isolate' CPU thread policy is in
effect, and this thread policy can only be used if using the 'dedicated'
CPU policy is used. As such, we don't need to worry about overhead if
either the 'share' CPU thread policy or the 'shared' CPU policy is in
effect.

Part of blueprint use-pcpu-and-vcpu-in-one-instance

Change-Id: Ib863b6a8d3bd3d2848fabd777fc1ac3387a7dd42
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>