libvirt: Change UEFI check to handle AArch64 better
Nova assumes that images run on UEFI instances will have
'hw_firmware_type' property set. This is wrong assumption because UEFI
is the default on AArch64, meaning images do not need to have this set.
This results in a failure to remove the instance, with libvirt raising
the following error:

  Instance failed to spawn: libvirt.libvirtError: Requested operation
  is not valid: cannot undefine domain with nvram

Resolve this by checking *both* the image metadata property and the
machine arch.

Change-Id: I2956fe2e3582c36d1c52a7e3becde1dacd9d41f0
Closes-bug: #1868464