Provide the parent pGPU when creating a new vGPU
Now that all compute node Resource Providers are reshaped (since Stein),
we know that each of them have children RPs if they have pGPUs.

That's why we can now pass the parent PCI ID (for the pGPU) when wanting
to create a new vGPU (ie. a mdev for libvirt).

NOTE(sbauza) : The current reshape functional test for VGPUs simulates
an old compute (pre-Stein) by creating by hand VGPU inventories on the
root RP and then creating the instances to the Nova API for making sure
that the guest XML will be written. This is OK but requires the
_allocate_mdev() code to be able to support non-reshaped VGPU
inventories.
Given moving to grenade checking the VGPU reshape testing wouldn't be
trivial, we accept to still provide support for non-nested RPs in this
specific _allocate_mdevs() libvirt. This sucks a bit, but we need to
keep it for a while and that doesn't harm.

Partially-Implements: blueprint vgpu-multiple-types

Change-Id: I6a1cebb6676b20b76e451fc3bde5ea54bbf16ff7